validation put into separate file to make code more clear
